ultimos testes

+++ entrypoint.sh
@@ -3,9 +3,12 @@
 # Garante que o script pare se algum comando falhar
 set -e
 
-# Limpa o cache de DI do Hyperf para evitar problemas em desenvolvimento
-echo "Limpando cache do container..."
-rm -rf /app/runtime/container/*
+# Em ambiente de desenvolvimento, limpa o cache do container para refletir novas alterações.
+# Em produção, essa etapa é pulada para uma inicialização mais rápida.
+if [ "${APP_ENV}" = "dev" ]; then
+    echo "Ambiente de desenvolvimento detectado. Limpando cache do container..."
+    rm -rf /app/runtime/container/*
+fi
 
 # Inicia a aplicação Hyperf
 exec php /app/bin/hyperf.php start
